DEV: Move existing clipboardCopy into utilities
