+++ app/assets/javascripts/discourse/app/initializers/copy-codeblocks.js
@@ -1,60 +1,9 @@
 import { cancel, later } from "@ember/runloop";
 import I18n from "I18n";
-import { Promise } from "rsvp";
 import { guidFor } from "@ember/object/internals";
+import { clipboardCopy } from "discourse/lib/utilities";
 import { iconHTML } from "discourse-common/lib/icon-library";
 import { withPluginApi } from "discourse/lib/plugin-api";
-
-// http://github.com/feross/clipboard-copy
-function clipboardCopy(text) {
-  // Use the Async Clipboard API when available.
-  // Requires a secure browsing context (i.e. HTTPS)
-  if (navigator.clipboard) {
-    return navigator.clipboard.writeText(text).catch(function (err) {
-      throw err !== undefined
-        ? err
-        : new DOMException("The request is not allowed", "NotAllowedError");
-    });
-  }
-
-  // ...Otherwise, use document.execCommand() fallback
-
-  // Put the text to copy into a <span>
-  const span = document.createElement("span");
-  span.textContent = text;
-
-  // Preserve consecutive spaces and newlines
-  span.style.whiteSpace = "pre";
-
-  // Add the <span> to the page
-  document.body.appendChild(span);
-
-  // Make a selection object representing the range of text selected by the user
-  const selection = window.getSelection();
-  const range = window.document.createRange();
-  selection.removeAllRanges();
-  range.selectNode(span);
-  selection.addRange(range);
-
-  // Copy text to the clipboard
-  let success = false;
-  try {
-    success = window.document.execCommand("copy");
-  } catch (err) {
-    // eslint-disable-next-line no-console
-    console.log("error", err);
-  }
-
-  // Cleanup
-  selection.removeAllRanges();
-  window.document.body.removeChild(span);
-
-  return success
-    ? Promise.resolve()
-    : Promise.reject(
-        new DOMException("The request is not allowed", "NotAllowedError")
-      );
-}
 
 let _copyCodeblocksClickHandlers = {};
 let _fadeCopyCodeblocksRunners = {};
